+++ workspace/solutions/rust/src/s0/s0/s0/s3/mod.rs
@@ -1,0 +1,5 @@
+pub mod s_0003_00;
+pub mod s_0003_01;
+pub mod s_0003_test;
+
+pub type TSolution = fn(String) -> i32;